requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.8.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
duckduckgo-search>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
